Tested bot but not fully started calling validate_user_input in handlers

+++ bot/schemas/variants/receivingvarquantity.py
@@ -1,3 +1,4 @@
+from decimal import Decimal
 from typing import Annotated
 
 from pydantic import BaseModel, StringConstraints
@@ -7,5 +8,5 @@ class ReceivingVarQuantity(BaseModel):
     quantity : int
     parent_id : int
     var_name : Annotated[str, StringConstraints(strip_whitespace = True, min_length = 1)]
-    var_price : float
+    var_price : Decimal
     admin_id : int
